requirements.txt
# Core ML Framework
torch>=2.0.0
torchvision>=0.15.0

# Carbon Tracking
codecarbon>=2.3.0

# Experiment Tracking & Logging
mlflow>=2.8.0
wandb>=0.15.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration Management
pyyaml>=6.0
argparse

# Utilities
tqdm>=4.65.0
scikit-learn>=1.3.0

# Web Dashboard
flask>=3.0.0
flask-cors>=4.0.0

# Optional: Alternative Dashboard
streamlit>=1.28.0

